Remove:login 타입 삭제
